refactor(core/app/advanced_chat): Update TTS publisher import path.
