+++ api/core/base/tts/__init__.py
@@ -1,0 +1,6 @@
+from core.base.tts.app_generator_tts_publisher import AppGeneratorTTSPublisher, AudioTrunk
+
+__all__ = [
+    "AppGeneratorTTSPublisher",
+    "AudioTrunk",
+]
